+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/StatesId.java
@@ -4,4 +4,6 @@ public interface StatesId {
     int INTRO = 0;
     int GAME = 1;
     int GAME_OVER = 2;
+
+    int MAIN_MENU = 10;
 }