+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/options/SimpleMenuOption.java
@@ -1,0 +1,8 @@
+package com.github.mjkuranda.spaceadventure2.states.menus.options;
+
+public class SimpleMenuOption extends MenuOption {
+
+    public SimpleMenuOption(String title) {
+        super(title);
+    }
+}