+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2023-07-26
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Menu, MenuOption and MenuOptionEvent system.
+- [@mjkuranda](https://github.com/mjkuranda): Main menu as first game state.
+- [@mjkuranda](https://github.com/mjkuranda): Simple start and exit options.
+
 ## [0.11.0] - 2023-07-11
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Image loading.